fix the browser redirect for long ips
commit 9691e1c "Made device url point to the device." uses
snprintf to create the webserver url (on port 80) to redirect
incoming traffic on the presentation url to. Since it uses
urlmax - 1 as buffer size, instead of urlmax as expected by
snprintf. If the ip is e.g. 192.168.100.123 it would silently
redirect to 192.168.100.12.

While the argument could be fixed for this case, it will still
not work when binding to an ipv6 interface. So just use dynamic
memory. (and yes there is no check if that succeeds, glibc
exits on malloc failure).

while at it, make the callback static, fix whitespace.